+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+ -  Ability to create HWMs
  -  Dropdown functionality to state and event filters on map page
  
 ### Changed